requirements.txt
Django==5.0.2
djangorestframework==3.14.0
mysqlclient==2.2.4
Pillow==10.2.0
python-decouple==3.8
gunicorn==21.2.0
django-rest-swagger==2.2.0
drf-yasg==1.21.7
